modify the component
